requirements/test.txt
-r base.txt
selenium==2.39.0
coverage==3.6
django-discover-runner==0.2.2
